Changed MapObject to use tile coordinates instead of pixel coordinates
Using tile coordinates makes sure we can resize an ObjectGroup without
it needing to be part of a map. Objects will also be able to adapt when
the tile size changes.

The position and size of objects is still saved in pixels.

+++ src/mapobject.cpp
@@ -24,7 +24,8 @@
 using namespace Tiled;
 
 MapObject::MapObject(const QString &name, const QString &type,
-                     int x, int y, int width, int height):
+                     qreal x, qreal y,
+                     qreal width, qreal height):
     mName(name),
     mPos(x, y),
     mSize(width, height),
